Update 060_forget.rst
Added overview description of difference between keep and keep-within. Also information that those not marked as keep will be deleted. It was not previously clear to me it function like a marking algorithm marking stuff to keep by different rules and delete the rest. I hope it is more clear with the short description I added.