Don't use locks for counting stats
Counting statistics uses locks each time an archiver needs to update
stats.

Use lockless stats counting, using locks only for measuring network
speeed (this bit is stateful and can't trivially be made lockless).

To be frank, I was curious if this was even feasible, and went out to
implement it mostly to know if it was. Benchmarks show a mere 1.6%
improvement in performance, with no real additional algorithmic
complexity for it.

Depends-On: https://github.com/restic/restic/pull/5764