restic backup: limit maximum size of repository during backup
A new option --max-repo-size is introduced and then passed on to
archiver/Snapshot().
Snapshot() monitors the current size of the repository by checking
arch.repoSize + int64(arch.summary.ItemStats.DataSizeInRepo + arch.summary.ItemStats.TreeSizeInRepo)
when a file gets saved. If the maximum is exceeded, a fatal error wil be triggered.

Currently, the snapshot gets aborted and there will be unreferenced packfiles in the
repository.