rewrite: restuctured code to use func walker.NodeKeepEmptyDirectoryFunc
NodeKeepEmptyDirectoryFunc will be called when a subdirectory found to be empty
to decide how to proceed.
NodeKeepEmptyDirectoryFunc is calling the --include pattern filter functions and
returns true if the empty subdirectory is to be kept, otherwise skipped.
If the subdirectory is to be discarded, it will return an empty nodeID i
with not error flag set. This will be picked up by the caller and processed it
accordingly.

This function is fed into NewSnapshotSizeRewriter.