restic backup --max-repo-size
Implementation of limiting the repository size during the backup process.

Introduce option --max-repo-size, which is converted to field
gopts.RepoSizeMax, and handed over to repository.New() in cmd/restic/global.go.

The command prints the current size in Verbosef mode at the beginning and
also at the end, when the repository size has been exceeded.