restic copy - add more status counters
Added struct CopyCounters which is modeled along archiver.Summary. Some of these
fields do not make sense in the context of the copy command, but they have been left,
so backup.ProgressPrinter() can be used to print the status information.

Added createSnapshotSummary() to create a restic.SnapshotSummary and save the new snapshot to the target
repository. Finally, progressReporter.Finish() will be used to create the status information either
in JSON or in text format.

cmd_copy_integration_test has be modified to accept the additional term parameter.