restic backup - new option `--max-repo-size`
cmd_backup will check for option `--max-repo-size`. If it is set, it will transfer its value
to GlobalOptions.RepoSizeMax, so `repository.New` can transfer it to `repo.opts.RepoSizeMax`.

Once the repository is open and `--max-repo-size` is set, the current repository size gets
calculated via `repo.CurrentRepositorySize()` and compared with the set maximum.
If this maxi,um is already exceeded, the backup process ends in error message
"repository maximum size already exceeded".

At the end of the run, repo.MaxCapacityExceeded() is checked and if set,
the message will be created in text mode output

=========================================
repository maximum size has been exceeded
Current repository size is xxx GiB
=========================================

The run is the terminated with the error message
"repository maximum size has been exceeded".