snapshot filtering: add --newer-than, --older-than
and --relative-to to filtering options for MultiSnapshotFilters

Introduce concept of 'DurationTime' which contains state and can hold
one of time.Time, restic.Duration or a snapshor string. All of these will
eventually be converted to a time.Time, so they can be used for time
related filter comparisons.

A 'DurationTime' will be defined for all of `--relative-to`, `--newer-than` and
`--older-than` and added to the restic.SnapshotFilter.

Changes:
internal/restic/snapshot_find.go: define `DurationTimeState`, its constants and
the functions for DurationTime(Set, String, Type, Empty).

cmd/restic/find.go: add the three filter options.

Addititon: changelog/unreleased/issue-5193