git prune - chose pruning based on packfile sizes
Additions: prune.go: created mapping of ``packsFromPackfiles`` so that
the information on all packfiles can be re-used.

Added function ``smallSizeEvaluation``which sorts the packfiles sizes and
then picks a percentile value from the the sorted list. A scale factor
will be apllied. If ``targetPackSize`` is smaller than the value returned
from this function, it will be replaced.

Added a test ``TestPrunePercentile`` in prune_test.go