Implement `backup --add-auto-tags` flag
When set, this causes snapshots created during a `backup` operation to
be tagged with a "__auto:restic:errors:archiver" tag in case there were
any handled errors during the operation, such as unreadable or
disappearing files.

This should make it easier to filter for potentially incomplete backups
with the `snapshots` command.