restic packfilelist
Command to show packfiles which belong to a snapshot or list of snapshots.

packfilelist gathers all snapshots from the the standard restic.SnapshotFilter,
collects the trees which belong to these snapshots and calls restic.FindUsedBlobs
to collect all blobs participating in these snapshots. These blobs get translated to
their containing packfiles and these lists are printed out either as standard text
of in JSON format. The `--detail` options allows for more insights to be looked at.