--skip-if-unchanged skips backup based on targets
Previously, when using the --skip-if-unchanged flag for a backup, restic would only check if the root tree ID remained the same.
This would mean if changes were made in unrelated paths, for example in the parent directories, the backup would still be made.
If this flag is enabled, restic now goes through each target and checks if the specified base targets have changed.